# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Insurance.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ClaimsOfficer/start-review": {
      "post": {
        "tags": [
          "ClaimsOfficer"
        ],
        "parameters": [
          {
            "name": "claimId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ClaimsOfficer/review": {
      "post": {
        "tags": [
          "ClaimsOfficer"
        ],
        "parameters": [
          {
            "name": "claimId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "approve",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ClaimsOfficer/settle": {
      "post": {
        "tags": [
          "ClaimsOfficer"
        ],
        "parameters": [
          {
            "name": "claimId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter 'Bearer' [space] and your token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}